+++ netdef/testutils.py
@@ -2,10 +2,31 @@ import datetime
 import inspect
 import os
 from unittest.mock import Mock
-
 from netdef.Engines.expression.Expression import Expression
 from netdef.Rules.utils import get_module_from_string
 from netdef.Sources.BaseSource import BaseSource, StatusCode
+from netdef.Shared.Shared import Shared
+from netdef.Shared import SharedConfig
+
+
+class MockShared(Shared):
+    def __init__(self, config_string=""):
+        identifier=""
+        install_path=None
+        proj_path=None
+        read_from_files = False
+
+        if "[general]" not in config_string:
+            config_string = "[general]\nversion=1\n{}".format(config_string)
+
+        default_config_string=config_string
+
+        self.config = SharedConfig.Config(
+            identifier, install_path, proj_path, default_config_string, read_from_files
+        )
+        self.sources = Mock()
+        self.expressions = Mock()
+        self.restart_on_exit = False
 
 
 class MockSource: